Move remaining CLI common helpers out of main
`_auto_output` and `output_json` are shared CLI helpers, not command-dispatch
logic. Move them into `mcp_video.cli.common` so `__main__.py` continues to
shrink before the larger handler split.

Constraint: Preserve JSON output behavior and generated output path behavior
Rejected: Move dispatch handlers in this change | keep this slice focused and easy to verify
Confidence: high
Scope-risk: narrow
Tested: python3 -m pytest tests/test_cli.py tests/test_doctor.py -q --tb=short; python3 -m pytest tests/test_cli.py::TestCLIError::test_invalid_file_outputs_json_error tests/test_cli.py tests/test_doctor.py -q --tb=short; python3 -m mcp_video --version; python3 -m mcp_video doctor --json; python3 -m ruff check mcp_video/__main__.py mcp_video/cli tests/test_cli.py; python3 -m ruff format --check mcp_video/__main__.py mcp_video/cli
Not-tested: full non-slow suite after small CLI helper cleanup
